requirements.txt
streamlit==1.40.2
google-generativeai==0.8.3
pydantic==2.12.5
